Switch to explicit relative imports as required in Python 3.x
The import of ``fields`` is an implicit relative import that is
no longer supported in Python 3.x. It has to be replaced with
``.fields`` to identify a relative import in the local package.

I've also added ``from __future__ import absolute_import`` to
force the same explicit behaviour in Python 2.x. This is
explained in more detail here:

https://wiki.python.org/moin/PortingPythonToPy3k#relative-imports